bug fixes & cleanup for deleting connect server info
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@4743 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/connect/org.wcs.smart.connect.dataqueue/src/org/wcs/smart/connect/dataqueue/LoginHandler.java
@@ -217,19 +217,21 @@ public class LoginHandler implements ILoginHandler {
 		if (files == null) return;
 		
 		Path p = FileSystems.getDefault().getPath(SmartDB.getCurrentConservationArea().getFileDataStoreLocation(), ConnectDataQueuePlugin.DATA_QUEUE_DIR);
-		Path root = FileSystems.getDefault().getPath(SmartContext.INSTANCE.getFilestoreLocation());
-		try(DirectoryStream<Path> systemfiles = Files.newDirectoryStream(p)){
-			for (Path temp : systemfiles){
-				if (!files.contains(root.relativize(temp).toString())){
-					try{
-						Files.deleteIfExists(temp);
-					}catch (Exception ex){
-						ConnectDataQueuePlugin.log(ex.getMessage(), ex);
+		if (Files.exists(p)){
+			Path root = FileSystems.getDefault().getPath(SmartContext.INSTANCE.getFilestoreLocation());
+			try(DirectoryStream<Path> systemfiles = Files.newDirectoryStream(p)){
+				for (Path temp : systemfiles){
+					if (!files.contains(root.relativize(temp).toString())){
+						try{
+							Files.deleteIfExists(temp);
+						}catch (Exception ex){
+							ConnectDataQueuePlugin.log(ex.getMessage(), ex);
+						}
 					}
 				}
+			}catch(Exception ex){
+				ConnectDataQueuePlugin.log(ex.getMessage(), ex);
 			}
-		}catch(Exception ex){
-			ConnectDataQueuePlugin.log(ex.getMessage(), ex);
 		}
 	}
 }
